fix: address Copilot review on PR #18
- src/composition.py:60 — drop dtype=torch.float16 hard-code; missing-key
  zero deltas now match the model parameter's dtype as the docstring
  promised. Prevents fp16/fp32 mixing during functional_call.
- src/clip_backends.py — replace global set_verbosity_error during model
  load with a targeted logging.Filter that drops only the cosmetic
  `position_ids` UNEXPECTED-keys row. Other warnings (real load
  failures, missing-keys reports) still surface.
- scripts/run_full_experiment_suite.sh — make CLIP_MODELS / OPENCLIP_MODELS
  env-overridable (`${CLIP_MODELS:-ViT-B-32}`) so the matrix can be
  widened without editing the script (e.g. `CLIP_MODELS="ViT-B-32 ViT-B-16
  ViT-L-14" ./scripts/run_full_experiment_suite.sh`).

+++ src/composition.py
@@ -57,7 +57,7 @@ def _align_task_vectors_by_name(task_vectors, params, param_names, context=""):
         for name, p in zip(param_names, params):
             v = tv.vector.get(name)
             if v is None:
-                deltas.append(torch.zeros_like(p, dtype=torch.float16))
+                deltas.append(torch.zeros_like(p))
             else:
                 if v.shape != p.shape:
                     raise RuntimeError(
